+++ tests/test_component.py
@@ -1533,7 +1533,7 @@ class TestComponentHook:
 
             def on_render(self, context: Context, template: Optional[Template]):
                 calls.append("slotted__on_render_pre")
-                _html, _error = yield template.render(context)  # type: ignore[union-attr]
+                _html, _error = yield lambda: template.render(context)  # type: ignore[union-attr]
 
                 calls.append("slotted__on_render_post")
 
@@ -1566,7 +1566,7 @@ class TestComponentHook:
                 if template is None:
                     yield None
                 else:
-                    _html, _error = yield template.render(context)
+                    _html, _error = yield lambda: template.render(context)
 
                 calls.append("inner__on_render_post")
 
@@ -1600,7 +1600,7 @@ class TestComponentHook:
 
             def on_render(self, context: Context, template: Optional[Template]):
                 calls.append("middle__on_render_pre")
-                _html, _error = yield template.render(context)  # type: ignore[union-attr]
+                _html, _error = yield lambda: template.render(context)  # type: ignore[union-attr]
 
                 calls.append("middle__on_render_post")
 
@@ -1632,7 +1632,7 @@ class TestComponentHook:
 
             def on_render(self, context: Context, template: Optional[Template]):
                 calls.append("outer__on_render_pre")
-                _html, _error = yield template.render(context)  # type: ignore[union-attr]
+                _html, _error = yield lambda: template.render(context)  # type: ignore[union-attr]
 
                 calls.append("outer__on_render_post")
 
@@ -1752,7 +1752,7 @@ class TestComponentHook:
                 # Check we can modify entries set by other methods
                 context["from_on_before__edited1"] = context["from_on_before"] + " (on_render)"
 
-                _html, _error = yield template.render(context)
+                _html, _error = yield lambda: template.render(context)
 
                 context["from_on_render_post"] = "3"
 
@@ -1804,7 +1804,7 @@ class TestComponentHook:
             def on_render(self, context: Context, template: Template):
                 template.nodelist.append(TextNode("\n---\nFROM_ON_RENDER_PRE"))
 
-                _html, _error = yield template.render(context)
+                _html, _error = yield lambda: template.render(context)
 
                 template.nodelist.append(TextNode("\n---\nFROM_ON_RENDER_POST"))
 
@@ -1831,6 +1831,61 @@ class TestComponentHook:
             """,
         )
 
+    def test_lambda_yield(self):
+        class SimpleComponent(Component):
+            template: types.django_html = """
+                text
+            """
+
+            def on_render(self, context: Context, template: Template):
+                html, _error = yield lambda: template.render(context)
+                return html + "<p>Hello</p>"
+
+        rendered = SimpleComponent.render()
+        assertHTMLEqual(
+            rendered,
+            "text<p data-djc-id-ca1bc3e>Hello</p>",
+        )
+
+        # Works without lambda
+        class SimpleComponent2(SimpleComponent):
+            def on_render(self, context: Context, template: Template):
+                html, _error = yield template.render(context)
+                return html + "<p>Hello</p>"
+
+        rendered2 = SimpleComponent2.render()
+        assertHTMLEqual(
+            rendered2,
+            "text<p data-djc-id-ca1bc3f>Hello</p>",
+        )
+
+    def test_lambda_yield_error(self):
+        def broken_template():
+            raise ValueError("BROKEN")
+
+        class SimpleComponent(Component):
+            def on_render(self, context: Context, template: Template):
+                _html, error = yield lambda: broken_template()
+                error.args = ("ERROR MODIFIED",)
+
+        with pytest.raises(
+            ValueError, match=re.escape("An error occured while rendering components SimpleComponent:\nERROR MODIFIED")
+        ):
+            SimpleComponent.render()
+
+        # Does NOT work without lambda
+        class SimpleComponent2(SimpleComponent):
+            def on_render(self, context: Context, template: Template):
+                # This raises an error instead of capturing it,
+                # so we never get to modifying the error.
+                _html, error = yield broken_template()
+                error.args = ("ERROR MODIFIED",)
+
+        with pytest.raises(
+            ValueError, match=re.escape("An error occured while rendering components SimpleComponent2:\nBROKEN")
+        ):
+            SimpleComponent2.render()
+
     def test_on_render_no_yield(self):
         class SimpleComponent(Component):
             template: types.django_html = """
@@ -1852,7 +1907,7 @@ class TestComponentHook:
             """
 
             def on_render(self, context: Context, template: Template):
-                _html, error = yield template.render(context)
+                _html, error = yield lambda: template.render(context)
 
                 raise error from None  # Re-raise original error
 
@@ -1879,15 +1934,15 @@ class TestComponentHook:
                 assert template is not None
 
                 with context.push({"case": 1}):
-                    html1, error1 = yield template.render(context)
+                    html1, error1 = yield lambda: template.render(context)
                     results.append((html1, error1))
 
                 with context.push({"case": 2}):
-                    html2, error2 = yield template.render(context)
+                    html2, error2 = yield lambda: template.render(context)
                     results.append((html2.strip(), error2))
 
                 with context.push({"case": 3}):
-                    html3, error3 = yield template.render(context)
+                    html3, error3 = yield lambda: template.render(context)
                     results.append((html3.strip(), error3))
 
                 html4, error4 = yield "<div>Other result</div>"
@@ -1990,7 +2045,7 @@ class TestComponentHook:
                         if template is None:
                             yield None
                         else:
-                            _html, _error = yield template.render(context)
+                            _html, _error = yield lambda: template.render(context)
                         return None  # noqa: PLR1711
 
             elif action == "no_return":
@@ -2000,7 +2055,7 @@ class TestComponentHook:
                         if template is None:
                             yield None
                         else:
-                            _html, _error = yield template.render(context)
+                            _html, _error = yield lambda: template.render(context)
 
             elif action == "raise_error":
 
@@ -2009,7 +2064,7 @@ class TestComponentHook:
                         if template is None:
                             yield None
                         else:
-                            _html, _error = yield template.render(context)
+                            _html, _error = yield lambda: template.render(context)
                         raise ValueError("ERROR_FROM_ON_RENDER")
 
             elif action == "return_html":
@@ -2019,7 +2074,7 @@ class TestComponentHook:
                         if template is None:
                             yield None
                         else:
-                            _html, _error = yield template.render(context)
+                            _html, _error = yield lambda: template.render(context)
                         return "HTML_FROM_ON_RENDER"
 
             else: